chore: move Filter to rpc-types
