+++ crates/rpc/rpc-types/src/eth/pubsub.rs
@@ -1,7 +1,7 @@
 //! Ethereum types for pub-sub
 
-use crate::{Log, RichHeader};
-use reth_primitives::{filter::Filter, H256};
+use crate::{eth::Filter, Log, RichHeader};
+use reth_primitives::H256;
 use serde::{de::Error, Deserialize, Deserializer, Serialize, Serializer};
 
 /// Subscription result.
